+++ packages/core/src/render3/util/view_traversal_utils.ts
@@ -13,17 +13,8 @@ import {LContainer} from '../interfaces/container';
 import {isLContainer, isLView} from '../interfaces/type_checks';
 import {CHILD_HEAD, CONTEXT, FLAGS, LView, LViewFlags, NEXT, PARENT} from '../interfaces/view';
 
+import {getLViewParent} from './view_utils';
 
-/**
- * Gets the parent LView of the passed LView, if the PARENT is an LContainer, will get the parent of
- * that LContainer, which is an LView
- * @param lView the lView whose parent to get
- */
-export function getLViewParent(lView: LView): LView|null {
-  ngDevMode && assertLView(lView);
-  const parent = lView[PARENT];
-  return isLContainer(parent) ? parent[PARENT]! : parent;
-}
 
 /**
  * Retrieve the root view from any component or `LView` by walking the parent `LView` until
